requirementes.txt
psycopg2-binary
python-dotenv
google-generativeai
openmeteo-requests
pandas
requests-cache
retry-requests